+++ src/crd/mod.rs
@@ -1,3 +1,4 @@
 pub mod gordo;
 pub mod model;
 pub mod pod;
+pub mod argo;
